Moved rqtl mapping code out of marker_regression and into its own file.
Later we will want to run all mapping methods outside of the web server, though
